+++ src/assets/navigation/noteIcon.png
