+++ src/assets/navigation/searchIcon.png
